requirements-api.txt
# Video Generator API 依赖

# Web 框架
fastapi>=0.80.0
uvicorn>=0.18.0
python-multipart>=0.0.5

# 数据处理
pydantic>=1.10.0

# 视频处理
moviepy==1.0.3
ffmpeg-python>=0.2.0
pydub>=0.25.1
pysubs2>=0.2.4

# 图像处理
opencv-python>=4.8.0
pillow>=10.0.0

# 工具
loguru>=0.7.0
click>=8.1.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0

# AI 功能
edge-tts>=6.1.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0
